Перенос одномерной и двумерной выборки в файл R (Начало рефактора)
